Refactor core inlining logic into src/ir/inlining.h
